+++ src/resonaate/dynamics/integration_events/continuous_state_change_event.py
@@ -1,4 +1,6 @@
 """Defines the :class:`.ContinuousStateChangeEvent` class."""
+from __future__ import annotations
+
 # Standard Library Imports
 from abc import ABCMeta, abstractmethod