+++ src/resonaate/estimation/adaptive/__init__.py
@@ -1,1 +1,3 @@
 """Contains all classes and functions related to adaptive estimation."""
+
+from __future__ import annotations