+++ src/resonaate/tasking/__init__.py
@@ -1,1 +1,2 @@
 """Contains definitions of sensor tasking framework to be solved as a POMDP."""
+from __future__ import annotations